Made some fixes for giveaway and ScheduledTask

+++ src/main/java/Main/ScheduledTask.java
@@ -3,7 +3,6 @@ package Main;
 import java.util.Calendar;
 import java.util.Timer;
 import java.util.TimerTask;
-import net.dv8tion.jda.api.events.Event;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import java.util.concurrent.TimeUnit;
 import net.dv8tion.jda.api.events.ReadyEvent;
@@ -15,29 +14,25 @@ public class ScheduledTask extends ListenerAdapter {
     public void onReady(@NotNull ReadyEvent event) {
 
         Timer timer = new Timer();
-        Calendar date = Calendar.getInstance();
-        date.set(Calendar.HOUR, 7);
-        date.set(Calendar.MINUTE, 0);
-        date.set(Calendar.SECOND, 0);
-        date.set(Calendar.MILLISECOND, 0);
-        timer.schedule(
-                new TimerTask() {
-                    @Override
-                    public void run() {
-                        refreshVCText(event);
-                    }
-                },
-                date.getTime(),
-                1000 * 60 * 60 * 24
-        );
+        TimerTask tt = new TimerTask(){
+            public void run(){
+
+                Calendar cal = Calendar.getInstance(); //this is the method you should use, not the Date(), because it is desperated.
+                int hour = cal.get(Calendar.HOUR_OF_DAY);//get the hour number of the day, from 0 to 23
+
+                if(hour == 7){
+                    event.getJDA().getGuildById("433825343485247499").getTextChannelsByName("vc-text", true).get(0).createCopy().queue();
+
+                    try { TimeUnit.SECONDS.sleep(1);
+                    } catch (InterruptedException e) { }
+
+                    event.getJDA().getGuildById("433825343485247499").getTextChannelsByName("vc-text", true).get(0).delete().queue();
+
+                    event.getJDA().getGuildById("433825343485247499").getTextChannelById("582399042240118814").sendMessage("VC-Text purged").queue();
+                }
+            }
+        };
+        timer.schedule(tt, 1000, 1000*60);
     }
 
-    public void refreshVCText(Event event){
-        event.getJDA().getGuildById("433825343485247499").getTextChannelsByName("vc-text", true).get(0).createCopy().queue();
-
-        try { TimeUnit.SECONDS.sleep(1);
-        } catch (InterruptedException e) { }
-
-        event.getJDA().getGuildById("433825343485247499").getTextChannelsByName("vc-text", true).get(0).delete().queue();
-    }
 }
